requirements.txt
chromadb>=0.4.22
langchain>=0.1.0
langchain-community
langchain-ollama>=0.0.1
langchain-openai>=0.0.5
langchain-text-splitters>=0.0.1
langgraph>=0.0.10
numpy<2.0.0,>=1.24.0
pydantic>=2.6.0
python-dotenv>=1.0.0
unstructured[pdf]>=0.12.0
unstructured[docx]>=0.12.0
python-docx>=1.1.0
pdf2image>=1.17.0
pdfminer.six>=20231228
pillow>=10.2.0
tiktoken>=0.5.2
